Use MethodHandle to improve performance.
This updates the Handlers to use use MethodHandle where it is possible, this is the first iteration of this which passes our unit tests.
This uses some Black Magic by using Reflection to get access to a field that we should not have access to in the Java API.

This allows for caching of default methods now.
The @Static annotation is now required as it is used to use the proper MethodHandle.

From internal testing this improves the time it takes to invoke the method from the Tunnel